refactor: remove svg-to-png functionality
This commit removes all the functionality for converting SVGs to PNGs. This feature was broken with previous updates and there is no current use that requires this feature to be fixed
