+++ packageScripts/postinstall.js
@@ -1,0 +1,14 @@
+'use strict';
+
+const chalk = require('chalk');
+const pjson = require('../package.json');
+
+console.log(chalk.hex('#f26135')(`
+╭ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ──────────────────────────────╮
+
+        Thanks for installing the latest version
+                of `) + chalk.hex('#ffd200').bold(`Auro Icons v${pjson.version}.`) + chalk.hex('#f26135')(`
+
+╰─────────────────────────────── ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─╯
+`)
+);
