--- scripts/colorTokensConfig.json
@@ -1,20 +1,0 @@
-{
-  "source": ["./node_modules/@alaskaairux/orion-design-tokens/**/*.json"],
-  "platforms": {
-    "colorTokens": {
-      "transformGroup": "css",
-      "buildPath": "./dist/tokens/",
-      "files": [
-        {
-          "destination": "colorToken.css",
-          "format": "css/variables",
-          "filter": {
-            "attributes": {
-              "category": "color"
-            }
-          }
-        }
-      ]
-    }
-  }
-}
